Change of directory structure proposal
*Api keys*
I think they should be in one file at the root of the repository, like: `user_config.example.json` with binance and bybit. This file could be given to the bot in the command line.

*Live settings file*
The magic behind getting the same file name in the live settings distub me, it could also be given through the CLI.

*Makefile*

Makefile is good to give a unique entry point and let the file take care of the change through time.